+++ CHANGELOG.md
@@ -20,11 +20,11 @@
 - [11528](https://github.com/vegaprotocol/vega/issues/11528) - Added support for eligible entities reward. 
 - [11372](https://github.com/vegaprotocol/vega/issues/11372) - Support combined filters for the `AMM` API.
 - [11535](https://github.com/vegaprotocol/vega/issues/11535) - Added support for lottery rank distribution strategy.
-### 🐛 Fixes
 
+### 🐛 Fixes
 - [11521](https://github.com/vegaprotocol/vega/issues/11521) - Restore `AMM` position factor when loading from a snapshot.
 - [11526](https://github.com/vegaprotocol/vega/issues/11526) - `EstimateAMMBounds` now respects the market's decimal places.
-
+- [11540](https://github.com/vegaprotocol/vega/issues/11540) - Fix spam check for spots to use not double count quantum.
 
 ## 0.77.5