+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - [11486](https://github.com/vegaprotocol/vega/issues/11486) - `AMMs` can now be submitted on markets with more decimal places than asset decimal places.
 - [11540](https://github.com/vegaprotocol/vega/issues/11540) - Fix spam check for spots to use not double count quantum.
 - [11542](https://github.com/vegaprotocol/vega/issues/11542) - Fix non determinism in lottery ranking.
+- [11544](https://github.com/vegaprotocol/vega/issues/11544) - Fix empty candles stream.
 
 
 ## 0.77.5